+++ CHANGELOG.md
@@ -15,6 +15,9 @@ This document is based on Keep a Changelog 1.0.0
 - Changed some default values (it only affects newly created objects and new object parameters).
 - Irrelevant options (i.e that have no effect) in some objects are no longer available (timers, visualizers). In consequence they are no longer stored as well with these objects (as they are useless).
 
+### Removed
+- Removed background visual object that duplicates image/shape visual object. Existing background visual objects in older saves will be converted to image visual objects.
+
 ## 0.2.2-beta (2021-07-28)
 
 ### Fixed